+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 2024.6.6
 ### Breaking Changes
+#### App
+- `Nickvision::App::Aura` now disables move operations, as well as previous disabled copy operations
 #### Events
 - `Nickvision::Events::HandlerId` is now defined as `enum class HandlerId : size_t` instead of `using HandlerId = size_t`
 #### Helpers